chore: consolidate analysis worker
